% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_separator.R
\name{detect_separator}
\alias{detect_separator}
\title{Detect the Separator in a GO Term String}
\usage{
detect_separator(go_string)
}
\arguments{
\item{go_string}{A character string containing GO terms.}
}
\value{
A character value indicating the detected separator (either \code{","} or \code{";"}).
}
\description{
This function detects the separator used in a string of Gene Ontology (GO) terms. It checks whether the terms are separated by commas (\verb{,}), semicolons (\verb{;}), or if no valid separator is found.
}
\examples{
# Example: GO terms separated by commas
detect_separator("GO:0008150,GO:0003674,GO:0005575")

# Example: GO terms separated by semicolons
detect_separator("GO:0008150;GO:0003674;GO:0005575")

# This will throw an error:
\dontrun{
detect_separator("GO:0008150 GO:0003674 GO:0005575")
}
}
