% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_TopgoEnrich.R
\name{detect_separator}
\alias{detect_separator}
\title{TopGo enrichment for a gene list}
\usage{
detect_separator(go_string)
}
\arguments{
\item{genelist}{Character vector of differentially expressed genes.}

\item{background}{Data frame with two columns: gene names and GO IDs.}

\item{ontology}{Ontology to use for analysis (default "BP").}

\item{reduce_terms}{Logical, whether to reduce redundant GO terms (default FALSE).}

\item{topnode}{number of nodes to report (default:30)}
}
\value{
Data frame with enrichment results.
}
\description{
Function to perform enrichment analysis with TopGo for all ontologies.
}
