% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunEnr_topgo.R
\name{FunEnr_Topgo}
\alias{FunEnr_Topgo}
\title{Perform GO Enrichment Analysis with topGO}
\usage{
FunEnr_Topgo(
  genelist,
  background,
  ontology = "BP",
  topnode = 30,
  reduce_terms = FALSE
)
}
\arguments{
\item{genelist}{A character vector of differentially expressed genes to be tested for enrichment.}

\item{background}{A data frame with two columns: the first column contains gene names, and the second column contains corresponding GO IDs.}

\item{ontology}{A character string specifying the ontology to use for analysis. Valid options are "BP" (Biological Process), "MF" (Molecular Function), and "CC" (Cellular Component). Defaults to "BP".}

\item{topnode}{An integer specifying the number of top GO terms (nodes) to report. Defaults to 30.}

\item{reduce_terms}{Logical, whether to reduce redundant GO terms in the results. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the enrichment analysis results, including GO terms, p-values, and scores.
}
\description{
This function performs Gene Ontology (GO) enrichment analysis using the \code{topGO} package for all GO ontologies (BP, MF, CC). It can optionally reduce redundant GO terms.
}
\details{
This function uses the \code{topGO} package to perform the analysis. It takes a gene list, compares it against the background GO terms, and identifies enriched GO categories based on the given ontology.
}
